Add `--init` to the Docker cmdline
Doing this has been discussed by folks running their own Docker miners. The typical reasons for not using tini (the --init flag) are if the code in the container is doing signal handling itself, if it is reaping zombies itself, or something else odd to PID 1 or the init system. From what I can tell, the validator/miner code is not doing those things, so I think adding `--init` to the recommended cmdline is useful.