Better representation of the acceptable data types
This change changes the types in the type column for body parameters. I think it is important to explain what kind of data is accepted in the api, especially the altitude field. I spent quite a long time trying to figure out why my data was not accepted (though not consistently) even though all my data looked correct. Turns out that altitude needs to be an integer and the lgt92 reported it as a float and an integer if there was no decimal points hence the data was accepted occasionally.