Chain following is done -- remove descriptor limit (#1024)
The directive to artificially expand the number of file descriptors available to the miner image is no longer needed. It is an artifact of earlier days, when each miner had to follow the blockchain, block for block. With the current hybrid approach and future Light Hotspot approach this is no longer necessary.