Fix reward claiming example.
The process of claiming rewards requires at least one call to the
distributor oracle to ask it to countersign the rewards transaction.
In doing so, the oracle will fill in the "recent block hash" for the
transaction and provide its signature over the same. This means
that the client code which uses this transaction must leave the
recent block hash alone when signing and submitting the transaction
to the chain.

`provider.sendAndConfirm()`, however, implicitly resets the recent
block hash and thus invalidates the transaction.

The fix here is to use a different submission method that leaves
the recent block hash as is.