Update lpn-km.mdx (#1454)
Corrected spelling.
Corrected weblink Product page because that was broken.
Added weblink Datasheet for their newer model.
The photo needs to be updated, but I couldn't find a good picture on their website.
The battery spec is different than from the datasheet. So I just left it like that for now.