Update digital-matter-lorawan-gps.mdx
Corrected spelling.
Deleted pictures converter cable. Both products use DM-Link v2 cable to do provisioning.
The myDevice/Cayenne info is actually not applicable, because Digital Matter is not supported under Helium in Cayenne.  Is it possible to add the Oyster3 to Cayenne under Helium?
I left the info in and updated it a bit, just in case we can get Oyster3 in.
The picture in the beginning needs to be updated as well to new Oyster3 and Yabby Edge pictures.