Updated sections in Governance
Updated committees.mdx
Updated hip.mdx
Updated edit-a-hip.mdx
Updated write-a-hip.mdx
Why is Edit-A-Hip and Write-A-Hip under faq in GitHub, but on the website it falls under heading Helium Improvement Proposals? For me it does not make sense to place it under faq. Not easy to find in GitHub to edit.