Update hnt-token.mdx (#1967)
Clarifies that the net emissions cap in HIP-20 was set at 1% of the epoch emissions *back then* and isn't adjusted with each halving.

Note: The paragraph "If less than 1,643.83561643 HNT is burned for DC within an epoch, the full amount burned will be
re-minted and added to the day's usual scheduled emissions for that day's epoch. However, if more
than 1,643.83561643 HNT is burned for DC within a single epoch, any HNT burn for DC over
1,643.83561643 will be permanently burned and removed from the max supply, while 1,643.83561643 will
be re-minted and distributed in scheduled emissions." isn't correct any more, now that we have the smoothing mechanism implemented. Flagging this, but have no capacity to update it myself.